API/requirements.txt
asgiref>=3.6.0<4
Django==4.2.7
pytz==2021.1
sqlparse==0.4.1
sqlalchemy==2.0.9
djangorestframework==3.14.0
djangorestframework-simplejwt
mysqlclient>=2.0
Pillow==10.1.0
celery
redis
django-redis==5.4.0
django-cors-headers==4.3.1


# pika==1.1.0